+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Release v0.1.4
 
     - Algorithm Gauss-Kronrod 15 (GK15) implemented for `integrate`.
+    - Improved `configure` to allow fine-grained GK15 configuration from Python.
+    - Improved `configure` to allow fine-grained Adaptive Simpson configuration from Python.
+    - Exposed `Value` comparison operators to Python.
 
 
 ## Release v0.1.3